+++ lib/Storage/CMakeLists.txt
@@ -2,8 +2,7 @@
 # This file is distributed under the MIT License. See LICENSE.md for details.
 #
 
-revng_add_library_internal(revngStorage SHARED StorageClient.cpp
-                           LocalStorageClient.cpp Path.cpp)
+revng_add_library_internal(revngStorage SHARED StorageClient.cpp Path.cpp)
 
 llvm_map_components_to_libnames(LLVM_LIBRARIES Core Support)
